remove section for Medium articles
Unfortunately it seems like there is no simple integration that does the job.
The free solutions do not work really well and have a bad performance when it
comes to responsibility. As we already have the link to our Medium blog, it has
been decided to remove this section for the moment and possibly target it later
again. After all it is not critical for a first release of the website.