convert paragraph lists into actual lists
There were some listings hidden in paragraph HTML elements with manual line
break, numeration etc. These have been converted to actual HTML lists. The
benefit is that we can now use the `outside` listing style, which correctly
indents multi-line text of a single list entry.

Unfortunately does the outside listing style lead to an issue where the numbers
are in-front of the whole block of text. Should mean it is not left aligned with
the headers and surrounding text. A quick solution (maybe there is
a better/correct one) was a manual left margin according to the font-size.