add media kit archive with link
Having the archive directly inside the repository is not 100% optimal.
I also decided against adapting the build process of the pages to somehow
include this asset too (does not work out of the box). As this is a simple link
in the page anyways, we can link to GitHub directly. Technically there is not
much difference.